requirements.txt
#sudo apt-get install python-poppler
requests
beautifulsoup4
Pillow
